requirements.txt
langchain
langchain-community
langchain-openai
python-dotenv
chromadb
openai
pypdf2
youtube-transcript-api
python-dotenv